test 코드 정리 및 불필요한 파일 삭제
