+++ src/test/dummyMemo.test.ts
@@ -1,4 +1,5 @@
 import path from 'path'
+import { Types } from 'mongoose'
 
 import dotenv from 'dotenv'
 
@@ -6,8 +7,8 @@ import { MemoModel } from '../model'
 import { HistoryModel } from '../model'
 import { PopulatedUser } from '../model/user'
 import { connectToDatabase } from '../utils/mongo'
-import { MEMOS } from '../dummy/dummy'
-import { TEMPTHISTORY } from '../dummy/dummy'
+import { MEMOS } from './dummy'
+import { TEMPTHISTORY } from './dummy'
 import axios from 'axios'
 
 jest.setTimeout(30000)
@@ -59,8 +60,8 @@ describe('make dummy memo and test', () => {
           method: 'post',
           data: {...TEMPTHISTORY,user:memo.user._id,category:memo.category._id}
         })
-        console.log(`데이터는 다음과 같다 -> ${TEMPTHISTORY}`)
-        expect(true).toStrictEqual(true)
+        const apiMemo = new Types.ObjectId(res?.data.newMemo.memo._id)
+        expect(apiMemo).toStrictEqual(memo._id)
       } catch (e) {
         console.error(e)
         return fail()
